+++ tests/test_utils.py
@@ -2,6 +2,7 @@ import csv
 from mipscripts import utils
 import pytest
 
+
 def test_to_snake_case():
     assert utils.to_snake_case("Sample.set") == "sample_set"
     assert utils.to_snake_case("Sample_set") == "sample_set"
@@ -9,6 +10,7 @@ def test_to_snake_case():
     assert utils.to_snake_case("sampleSet") == "sample_set"
     assert utils.to_snake_case("sample Set") == "sample_set"
     assert utils.to_snake_case("sample\nSet") == "sample_set"
+
 
 def test_header_to_snake_case(tmp_path):
     tmp_dir = tmp_path / "header"
@@ -58,3 +60,21 @@ def test_header_to_snake_case(tmp_path):
 
     # Check that headers match
     assert res_header == expect_header
+
+
+def test_duplicate_cols(tmp_path):
+    tmp_dir = tmp_path / "header"
+    tmp_dir.mkdir()
+    tmp_file = tmp_dir / "sample_list.tsv"
+
+    # Create initial header
+    init_header = ["sample_name", "sampleName"]
+
+    # Write header to file
+    with open(tmp_file, mode="w") as file:
+        file_content = csv.writer(file, delimiter="\t")
+        file_content.writerow(init_header)
+
+    # Convert header to snake case
+    with pytest.raises(SystemExit):
+        utils.header_to_snake_case(tmp_file, overwrite=True)